Run specs against Mongoid as well as ActiveRecord
Splits AR-specific setup out into conditionally-required files. Runs the
same specs against both ORMs. Right now, that is controlled by an ENV var:

    $ ORM=active_record rake spec
    $ ORM=mongoid rake spec
