+++ acts_as_tenant.gemspec
@@ -27,6 +27,7 @@ Gem::Specification.new do |s|
   s.add_development_dependency('rspec-rails')
   s.add_development_dependency('database_cleaner')
   s.add_development_dependency('sqlite3')
+  s.add_development_dependency('mongoid', '~> 4.0')
 
   s.add_development_dependency('sidekiq', '3.2.1')
 end
