+++ spec/mongoid.yml
@@ -1,0 +1,6 @@
+test:
+  sessions:
+    default:
+      database: acts_as_tenant_test
+      hosts:
+        - localhost:27017
