+++ spec/spec_helper.rb
@@ -1,27 +1,14 @@
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
 $LOAD_PATH.unshift(File.dirname(__FILE__))
 
-require 'rails/all'
+$orm = ENV["ORM"] || "active_record"
+require "#{$orm}_helper"
+
 require 'rspec/rails'
-require 'database_cleaner'
 require 'acts_as_tenant'
 
-config = YAML::load(IO.read(File.join(File.dirname(__FILE__), 'database.yml')))
-ActiveRecord::Base.logger = Logger.new(File.join(File.dirname(__FILE__), "debug.log"))
-ActiveRecord::Base.establish_connection(config[ENV['DB'] || 'sqlite'])
-
 RSpec.configure do |config|
-  config.before(:suite) do
-    DatabaseCleaner.strategy = :transaction
-    DatabaseCleaner.clean_with(:truncation)
-  end
-
-  config.before(:each) do
-    DatabaseCleaner.start
-  end
-
   config.after(:each) do
-    DatabaseCleaner.clean
     ActsAsTenant.current_tenant = nil
   end
